refactor(demodata): move redux side effects into thunks
